requirements.txt
APScheduler==3.9.1
arabic-reshaper==2.1.3
asgiref~=3.3.1
asn1crypto~=1.0.1
backports.zoneinfo==0.2.1
certifi~=2020.6.20
cffi~=1.12.3
charset-normalizer==2.1.0
click~=7.0
colorama~=0.4.1
cryptography~=2.9.2
cssselect2~=0.3.0
dj-database-url==1.0.0
Django~=3.0.6
django-crispy-forms~=1.9.2
django-filter~=2.2.0
future~=0.17.1
gunicorn~=20.0.4
html5lib~=1.0.1
idna~=2.6
lxml~=4.4.1
oscrypto==1.3.0
Pillow~=7.0.0
psycopg2~=2.8.6
psycopg2-binary==2.9.3
pycparser~=2.19
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
pytz~=2020.5
pytz-deprecation-shim==0.1.0.post0
PyYAML~=5.3.1
qrcode==7.3.1
reportlab~=3.5.53
requests~=2.24.0
schedule==1.1.0
six~=1.15.0
sqlparse~=0.4.1
svglib==1.4.1
tinycss2~=1.0.2
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3~=1.25.10
webencodings==0.5.1
xhtml2pdf~=0.2.5

defusedxml~=0.6.0
PyQt5~=5.12.3
ipython~=7.8.0
olefile~=0.46
pip~=21.3.1
attrs~=19.2.0
wheel~=0.34.2
docutils~=0.15.2
Sphinx~=2.2.0
tornado~=6.1
Jinja2~=2.10.3
keyring~=18.0.0
beautifulsoup4~=4.8.0
dask~=2.5.2
tblib~=1.6.0
geoip2~=4.1.0
numpy~=1.16.5
environ~=1.0
aiohttp~=3.6.2
gevent~=1.4.0
eventlet~=0.25.2
greenlet~=0.4.15
py~=1.8.0
chardet~=3.0.4
ply~=3.11
tables~=3.6.1
Pyphen~=0.9.5
setuptools~=41.2.0
Cython~=0.29.13
SQLAlchemy~=1.3.9
pytest~=5.2.1
bcrypt~=3.1.7
path~=13.1.0
python-dateutil~=2.8.1